Refactoring out the Edge struct
