+++ Assets/HoloToolkit/Boundary/Scripts/Edge.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 84b2e4678d87a0344b197bbc74ba1f3a
+timeCreated: 1530819263
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
